Move tso dispatching and stream processing related code to tso_request_dispatcher.go to improve code readability.
